+++ src/redux/configureStore.js
@@ -1,6 +1,7 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
 import thunk from 'redux-thunk';
+import { booksReducer, loadingReducer } from './books/reducers/books';
 
 import { booksReducer, loadingReducer } from './books/books';
